Bump Lodestar deps to 0.32.0, BLS/SSZ to latest
This updates Lodestar dependencies to version 0.32.0, and BLS/SSZ to the
latest version, using the command:
```
yarn upgrade -S @chainsafe --latest
```

Notably, this fixes an incompatibility when connecting to the
`/eth/v1/config/spec` REST endpoint of a non-Lodestar server.
This also changes the `/eth/v1/lightclient/init_proof/:blockRoot`
endpoint to be parameterized by root instead of epoch value, and
to return the proof for the given block's post-state, instead of
the epoch's boundary post-state.